+++ app/src/main/kotlin/swc/microservice/truck/entities/Position.kt
@@ -1,3 +1,3 @@
 package swc.microservice.truck.entities
 
-data class Position(val latitude: Long, val longitude: Long)
+data class Position(val latitude: Double, val longitude: Double)
